+++ src/module/feature/slash-commands/models/detected-clan.model.ts
@@ -1,16 +1,14 @@
 import { type Client, Colors, EmbedBuilder, type TextChannel } from 'discord.js';
-import type { WotApiModel } from '../../../shared/apis/wot-api/wot-api.model';
+import type { WotApi } from '../../../shared/apis/wot/wot.api';
+import { Api } from '../../../shared/decorators/injector/api-injector.decorator';
 import { LoggerInjector } from '../../../shared/decorators/injector/logger-injector.decorator';
-import { Singleton } from '../../../shared/decorators/injector/singleton-injector.decorator';
 import { Table } from '../../../shared/decorators/injector/table-injector.decorator';
 import { EmojiEnum } from '../../../shared/enums/emoji.enum';
 import type { ChannelsTable } from '../../../shared/tables/complexe-table/channels/channels.table';
-import type { FoldRecruitmentTable } from '../../../shared/tables/complexe-table/fold-recruitment/fold-recruitment.table';
-import type { WotApiTable } from '../../../shared/tables/complexe-table/wot-api/wot-api.table';
 import type { PotentialClansTable } from '../../../shared/tables/simple-table/potential-clans.table';
 import type { Logger } from '../../../shared/utils/logger';
+import { UrlUtil } from '../../../shared/utils/url.util';
 import { UserUtil } from '../../../shared/utils/user.util';
-import { FoldRecruitmentEnum } from '../../loops/enums/fold-recruitment.enum';
 
 /**
  * DetectedClanModel class responsible for managing and processing detected clans.
@@ -21,9 +19,7 @@ export class DetectedClanModel {
     private readonly logger: Logger;
     @Table('PotentialClans') private readonly potentialClanTable: PotentialClansTable;
     @Table('Channels') private readonly channelsTable: ChannelsTable;
-    @Table('FoldRecruitment') private readonly foldRecruitmentTable: FoldRecruitmentTable;
-    @Table('WotApi') private readonly wotApiTable: WotApiTable;
-    @Singleton('WotApi') private readonly wotApiModel: WotApiModel;
+    @Api('Wot') private readonly wotApi: WotApi;
     //endregion
 
     /**
@@ -37,24 +33,12 @@ export class DetectedClanModel {
     private channel: TextChannel;
 
     /**
-     * URL for accessing Wargaming clan data.
-     */
-    private wargamingUrl: string;
-
-    /**
-     * URL for accessing Wot Life clan data.
-     */
-    private wotLifeUrl: string;
-
-    /**
      * Initializes the model with necessary data.
      *
      * @param {Client} client - The client instance to use.
      */
     public async initialize(client: Client): Promise<void> {
         this.channel = await UserUtil.fetchChannelFromClient(client, await this.channelsTable.getFoldRecruitment());
-        this.wargamingUrl = await this.foldRecruitmentTable.getUrl('clan');
-        this.wotLifeUrl = await this.wotApiTable.getUrl('wot_life_clan');
     }
 
     /**
@@ -75,12 +59,12 @@ export class DetectedClanModel {
         let embed = this.buildEmbed();
         let numberOfFields = 0;
 
-        for (const clan of this.clans) {
-            const tag: string = (await this.wotApiModel.fetchClanDetails(clan)).data[clan].tag;
+        for (const clanId of this.clans) {
+            const tag: string = (await this.wotApi.clansInfo(clanId)).data[clanId].tag;
 
             embed.addFields({
                 name: `${tag}`,
-                value: `[WG](${this.wargamingUrl.replace(FoldRecruitmentEnum.CLAN_ID, String(clan))}) ${EmojiEnum.REDIRECTION} | [Wot Life](${this.wotLifeUrl.replace(FoldRecruitmentEnum.CLAN_ID, String(clan)).replace(FoldRecruitmentEnum.CLAN_NAME, tag)}) ${EmojiEnum.REDIRECTION}`,
+                value: `[WG](${UrlUtil.getWargamingClanUrl(clanId)}) ${EmojiEnum.REDIRECTION} | [Wot Life](${UrlUtil.getWotLifeClanUrl(tag, clanId)}) ${EmojiEnum.REDIRECTION}`,
                 inline: true,
             });