+++ src/module/shared/apis/wot/wot.api.ts
@@ -1,15 +1,6 @@
-import type { AxiosInstance } from 'axios';
 import { application_id_wot } from '../../../core/config.json';
-import { FoldRecruitmentEnum } from '../../../feature/loops/enums/fold-recruitment.enum';
 import { LoggerInjector } from '../../decorators/injector/logger-injector.decorator';
-import { Singleton } from '../../decorators/injector/singleton-injector.decorator';
-import { Table } from '../../decorators/injector/table-injector.decorator';
-import { EmojiEnum } from '../../enums/emoji.enum';
-import { TimeEnum } from '../../enums/time.enum';
-import type { WotApiTable } from '../../tables/complexe-table/wot-api/wot-api.table';
-import type { Logger } from '../../utils/logger';
-import type { WargamingErrorType, WargamingSuccessType } from './models/wargaming-api.type';
-import { WotApiConstants } from './models/wot-api.enum';
+import { ApiBase } from '../api.base';
 import type {
     ClanDetailsDto,
     ClanDetailsSuccess,
@@ -22,33 +13,26 @@ import type {
     TankopediaVehicle,
     TankopediaVehiclesSuccess,
 } from './models/wot-api.type';
+import type { WargamingErrorType, WargamingSuccessType } from './models/wot-base-api.type';
 
 @LoggerInjector
-export class WotApiModel {
-    /**
-     * The base url for the wot api
-     * @default https://api.worldoftanks.eu
-     */
-    private readonly WOT_API: string = 'https://api.worldoftanks.eu';
-
+export class WotApi extends ApiBase {
     /**
      * The maximum number of try when fetching vehicle's data
      * @default 5
      */
     private readonly maxNumberOfTry: number = 5;
 
-    //region INJECTABLE
-    private readonly logger: Logger;
-    @Singleton('Axios', TimeEnum.SECONDE * 30) private readonly axios: AxiosInstance;
-    @Table('WotApi') private readonly wotApi: WotApiTable;
-    //endregion
+    constructor() {
+        super('https://api.worldoftanks.eu');
+    }
 
     /**
-     * Fetches data from the Tankopedia API.
+     * Get the data of vehicles from teh tankopedia api
      *
-     * @param {number} [pageNumber] - The page number on the Tankopedia Api
+     * @param {number} [pageNumber] - The page number on the Tankopedia Api. This page number is link to a tank : page 1 in tier 10 will always be Kranvagn
      *
-     * @returns {Promise<TankopediaVehiclesSuccess>} - A promise that resolves with the tankopedia data on success.
+     * @returns {Promise<TankopediaVehiclesSuccess>} - A promise that resolves with the data of the vehicle on success.
      *
      * @example
      * const tankopediaData: TankopediaVehiclesSuccess = await instance.fetchTankopediaApi(2);
@@ -58,21 +42,22 @@ export class WotApiModel {
      * const tankopediaData: TankopediaVehiclesSuccess = await instance.fetchTankopediaApi();
      * console.log(tankopediaData); // { status: 'ok', meta: { count: 50, total: 50 }, data: { ... } }
      */
-    public async fetchTankopediaApi(pageNumber?: number): Promise<TankopediaVehiclesSuccess> {
-        let url = this.concatUrl(await this.wotApi.getUrl('trivia'));
+    public async tankopediaVehicles(pageNumber?: number): Promise<TankopediaVehiclesSuccess> {
+        const url: URL = this.createUrl(
+            '/wot/encyclopedia/vehicles/?tier=10&language=fr&fields=name%2C+images.big_icon%2C+default_profile.ammo.damage%2C+default_profile.ammo.type&limit=1&type=heavyTank%2C+AT-SPG%2C+mediumTank%2C+lightTank'
+        );
 
         if (pageNumber) {
-            url += `&page_no=${pageNumber}`;
+            this.addSearchParam(url, 'page_no', pageNumber);
         }
 
-        this.logger.debug(`${EmojiEnum.SOLDIER} Fetching tankopedia api with url {}`, url);
         return await this.getDataFromUrl<TankopediaVehiclesSuccess, TankopediaVehicle>(url);
     }
 
     /**
      * Fetches clan image data from the Wargaming API.
      *
-     * @param {string} name - The name of the clan.
+     * @param {string} clanName - The name of the clan.
      *
      * @returns {Promise<ClansSuccess>} - A promise that resolves with the clan image data on success.
      *
@@ -80,10 +65,9 @@ export class WotApiModel {
      * const clanImageData: ClansSuccess = await instance.fetchClanImage('clan_name');
      * console.log(clanImageData); // { status: 'ok', meta: { count: 1, total: 1 }, data: [{ emblems: { x64: { wot: 'clan_wot_emblem_url', portal: 'clan_portal_emblem_url' } } }] }
      */
-    public async fetchClanImage(name: string): Promise<ClansSuccess> {
-        const url: string = this.concatUrl((await this.wotApi.getUrl('image_url')).replace(WotApiConstants.CLAN_NAME, name));
-        this.logger.debug(`${EmojiEnum.SOLDIER} Fetching clans image of clan {} with url {}`, name, url);
-
+    public async clansList(clanName: string): Promise<ClansSuccess> {
+        const url: URL = this.createUrl('/wot/clans/list/?fields=emblems.x64');
+        this.addSearchParam(url, 'search', clanName);
         return await this.getDataFromUrl<ClansSuccess, ClansDto>(url);
     }
 
@@ -94,12 +78,9 @@ export class WotApiModel {
      *
      * @returns {Promise<PlayerPersonalDataSuccess>} - A Promise that resolves with the personal data of the player.
      */
-    public async fetchPlayerPersonalData(playerId: number): Promise<PlayerPersonalDataSuccess> {
-        const url: string = this.concatUrl(
-            (await this.wotApi.getUrl('player_personal_data')).replace(FoldRecruitmentEnum.PLAYER_ID, String(playerId))
-        );
-        this.logger.debug(`${EmojiEnum.SOLDIER} Fetching personal data with url {}`, url);
-
+    public async accountInfo(playerId: number): Promise<PlayerPersonalDataSuccess> {
+        const url: URL = this.createUrl('/wot/account/info/?account_id=PLAYER_ID&fields=clan_id');
+        this.addSearchParam(url, 'account_id', playerId);
         return await this.getDataFromUrl<PlayerPersonalDataSuccess, PlayerPersonalDto>(url);
     }
 
@@ -110,14 +91,9 @@ export class WotApiModel {
      *
      * @returns {Promise<PlayerDto>} - A Promise that resolves with the data of the player.
      */
-    public async fetchPlayerData(playerName: string): Promise<PlayerDataSuccess> {
-        const url: string =
-            this.WOT_API +
-            (await this.wotApi.getUrl('player_url'))
-                .replace(WotApiConstants.APPLICATION_ID, application_id_wot)
-                .replace(FoldRecruitmentEnum.PLAYER_NAME, String(playerName));
-        this.logger.debug(`${EmojiEnum.SOLDIER} Fetching player data with url {}`, url);
-
+    public async accountList(playerName: string): Promise<PlayerDataSuccess> {
+        const url: URL = this.createUrl('/wot/account/list/');
+        this.addSearchParam(url, 'search', playerName);
         return await this.getDataFromUrl<PlayerDataSuccess, PlayerDto>(url);
     }
 
@@ -134,30 +110,19 @@ export class WotApiModel {
      * const clanDetails = await fetchClanDetails(12345);
      * console.log(clanDetails); // { "12345": { tag: "TAG" } }
      */
-    public async fetchClanDetails(clanId: number): Promise<ClanDetailsSuccess> {
-        const url: string =
-            this.WOT_API +
-            (await this.wotApi.getUrl('clan_details'))
-                .replace(WotApiConstants.APPLICATION_ID, application_id_wot)
-                .replace(FoldRecruitmentEnum.CLAN_ID, String(clanId));
-        this.logger.debug(`${EmojiEnum.SOLDIER} Fetching clan details with url {}`, url);
-
+    public async clansInfo(clanId: number): Promise<ClanDetailsSuccess> {
+        const url: URL = this.createUrl('/wot/clans/info/');
+        this.addSearchParam(url, 'clan_id', clanId);
         return await this.getDataFromUrl<ClanDetailsSuccess, ClanDetailsDto>(url);
     }
 
     /**
-     * Concat the start of the WoT api url with the following api and replace the application ID with the value
-     *
-     * @param {string} url - The url to concat
-     *
-     * @return {string} - The url concat with the right value for the application id
-     *
-     * @example
-     * const url = this.concatUrl('/wot/api?application_id=APPLICATION_ID");
-     * console.log(url) // 'https://api.worldoftanks.eu/wot/api?application_id=123456789'
+     * @inheritdoc
      */
-    private concatUrl(url: string): string {
-        return this.WOT_API + url.replace(WotApiConstants.APPLICATION_ID, application_id_wot);
+    protected override createUrl(endpoint: string): URL {
+        const url: URL = super.createUrl(endpoint);
+        this.addSearchParam(url, 'application_id', application_id_wot);
+        return url;
     }
 
     /**
@@ -174,14 +139,12 @@ export class WotApiModel {
      * @template GSuccess - The type of the successful data.
      * @template GData - The generic type representing either WargamingSuccessType or WargamingErrorType.
      */
-    private async getDataFromUrl<GSuccess, GData extends WargamingSuccessType<object> | WargamingErrorType>(
-        url: string
-    ): Promise<GSuccess> {
+    private async getDataFromUrl<GSuccess, GData extends WargamingSuccessType<object> | WargamingErrorType>(url: URL): Promise<GSuccess> {
         let numberOfTry: number = 0;
         let data: GData;
 
         do {
-            data = (await this.axios.get(url)).data;
+            data = await this.getData(url);
         } while (!data && numberOfTry++ < this.maxNumberOfTry);
 
         if (!data || data.status === 'error') {