+++ src/module/shared/decorators/injector/models/singleton-injector.type.ts
@@ -1,4 +1,0 @@
-/**
- * All the singleton that can be injected with the decorator
- */
-export type SingletonDependence = 'Trivia' | 'Axios' | 'WotApi' | 'Database';