+++ src/module/shared/decorators/injector/table-injector.decorator.ts
@@ -1,4 +1,4 @@
-import type { Constructor } from './models/constructor.type';
+import type { Constructor } from './models/injector.type';
 
 let tableMap: {
     // Commons
@@ -14,7 +14,6 @@ let tableMap: {
     // Newsletter
     NewsWebsites: Constructor;
     BanWords: Constructor;
-    WotApi: Constructor;
     // Trivia game
     TriviaData: Constructor;
     Tanks: Constructor;
@@ -52,11 +51,10 @@ export function Table(
                 .BlacklistedPlayersTable,
             LeavingPlayers: require('../../tables/complexe-table/leaving-players/leaving-players.table').LeavingPlayersTable,
             PotentialClans: require('../../tables/simple-table/potential-clans.table').PotentialClansTable,
-            FoldRecruitment: require('../../tables/complexe-table/fold-recruitment/fold-recruitment.table').FoldRecruitmentTable,
+            FoldRecruitment: require('../../tables/simple-table/fold-recruitment.table').FoldRecruitmentTable,
             // Newsletter
             NewsWebsites: require('../../tables/complexe-table/news-websites/news-websites.table').NewsWebsitesTable,
             BanWords: require('../../tables/simple-table/ban-words.table').BanWordsTable,
-            WotApi: require('../../tables/complexe-table/wot-api/wot-api.table').WotApiTable,
             // Trivia game
             TriviaData: require('../../tables/complexe-table/trivia-data/trivia-data.table').TriviaDataTable,
             Tanks: require('../../tables/complexe-table/tanks/tanks.table').TanksTable,