+++ src/module/shared/tables/complexe-table/fold-recruitment/models/fold-recruitment.type.ts
@@ -1,1 +1,0 @@
-export type FoldRecruitmentUrl = 'newsfeed' | 'clan' | 'tomato' | 'wargaming' | 'wot_life';