+++ src/module/shared/tables/complexe-table/wot-api/models/wot-api.type.ts
@@ -1,1 +1,0 @@
-export type WotApi = 'image_url' | 'player_url' | 'player_personal_data' | 'trivia' | 'clan_details' | 'wot_life_clan';